compute all matches even in "perfect" happy path
The "perfect" happy path is for when there are at least
`visible_choices` strings with a perfect score. I only generated the
first `visible_choices` to avoid analyzing all possible matches. But
that's wrong, since matches should be sub-sorted by length. That means
that we always have to compute all matches in case a very short one is
at the end of the choice list.